% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvexCluster_wrapper.R
\name{convexADMM}
\alias{convexADMM}
\title{Title}
\usage{
convexADMM(
  X,
  Lambda,
  ix,
  M1,
  M2,
  s1,
  s2,
  w,
  gamma,
  nu,
  num_iter = 100,
  tol_abs = 1e-05,
  tol_rel = 1e-04
)
}
\arguments{
\item{tol_rel}{}
}
\value{

}
\description{
Title
}
