\name{ConvexClust-package}
\alias{ConvexClust-package}
\alias{ConvexClust}
\docType{package}
\title{
\packageTitle{ConvexClust}
}
\description{
\packageDescription{ConvexClust}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ConvexClust}
\packageIndices{ConvexClust}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ConvexClust}

Maintainer: \packageMaintainer{ConvexClust}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
